# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Swagger for withDoc",
    "description": "Open API docs with Swagger",
    "license": {
      "name": "MIT",
      "url": "http://localhost:3000"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Task Swagger",
      "description": "API for task"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/": {
      "get": {
        "tags": ["Main"],
        "summary": "Main Page Get",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#definitions/item"
            }
          }
        }
      },
      "post": {
        "tags": ["Main"],
        "summary": "Login Data Post",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Register",
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Register Success"
          },
          "404": {
            "description": "Failed. Bad Request"
          }
        }
      }
    },
    "/user/signup" : {
      "post": {
        "tags": ["User"],
        "summary": "SignUp new user",
        "parameters": [
          {
            "in":"body",
            "name": "body",
            "description": "Register",
            "schema":{
              "$ref":"#/definitions/new user"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Register Success"
          },
          "403": {
            "description": "Failed. Bad Request"
          }
        }
      }
    },
    "/user/signin" : {
      "post": {
        "tags": ["User"],
        "summary": "Login",
        "parameters": [
          {
            "in":"body",
            "name": "body",
            "description": "Register",
            "schema":{
              "$ref":"#/definitions/user login"
            }
          }
        ],
        "responses": {
          "201": {
            "type":"jwt",
            "token":"asdbaskdhbaskdjadkjasdnkasd"
          },
          "501": {
            "description": "does not exist email or invalid password"
          }
        }
      }
    },
    "/user/" : {
      "put": {
        "tags": ["User"],
        "summary": "Modify user password",
        "parameters": [
          {
            "in":"body",
            "name": "body",
            "description": "update",
            "schema":{
              "$ref":"#/definitions/user modify"
            }
          }
        ],
        "responses": {
          "201": "ok",
          "501": {
            "description": "does not exist email or invalid password"
          }
        }
      }
    },
    "/user" : {
      "delete": {
        "tags": ["User"],
        "summary": "delete user",
        "parameters": [
          {
            "in":"body",
            "name": "body",
            "description": "delete"
          }
        ],
        "responses": {
          "201": "ok",
          "501": {
            "description": "does not exist email or invalid password"
          }
        }
      }
    },
    "/document/all" : {
      "get": {
        "tags": ["Document"],
        "summary": "get documents",
        "parameters": [
          {
            "in":"body",
            "name": "body",
            "description": "get documents"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#definitions/get documents"
            }
          },
          "501": {
            "description": "does not exist email or invalid password"
          }
        }
      }
    },
    "/document" : {
      "post": {
        "tags": ["Document"],
        "summary": "post document",
        "parameters": [
          {
            "in":"body",
            "name": "body",
            "description": "create",
            "schema":{
              "$ref":"#/definitions/post document"
            }
          }
        ],
        "responses": {
          "201": "ok",
          "501": {
            "description": "does not exist email or invalid password"
          }
        }
      },
      "get": {
        "tags": ["Document"],
        "summary": "get specific document info",
        "parameters": [
          {
            "in":"body",
            "name": "body",
            "description": "get specific document info",
            "schema":{
              "$ref":"#/definitions/get specific doc req"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#definitions/get document"
            }
          },
          "501": {
            "description": "does not exist email or invalid password"
          }
        }
      }
    },
    "/document/" : {
      "delete": {
        "tags": ["Document"],
        "summary": "delete document",
        "parameters": [
          {
            "in":"body",
            "name": "body",
            "description": "delete document"
          }
        ],
        "responses": {
          "201": "ok",
          "501": {
            "description": "does not exist email or invalid password"
          }
        }
      }
    }
  },
  "definitions": {
    "get specific doc req":{
      "properties":{
          "docId": {"type":"string"}
        }
    },
    "get all doc req":{
      "properties":{
          "email": {"type":"string"}
        }
    },
    "get documents":{
      "properties":{
          "dataType": {"type":"string"},
          "docExpiryDate": {"type":"string"},
          "docName": {"type":"string"},
          "docPublishedDate": {"type":"date"},
          "docPublishedOrg": {"type":"string"},
          "docSerialNum": {"type":"string"},
          "docType": {"type":"string"},
          "id": {"type":"string"},
          "userId": {"type":"string"},
          "visaType": {"type":"string"}
        }
    },
    "get document":{
      "properties":{
          "dataType": {"type":"string"},
          "docExpiryDate": {"type":"string"},
          "docName": {"type":"string"},
          "docPublishedDate": {"type":"date"},
          "docPublishedOrg": {"type":"string"},
          "docSerialNum": {"type":"string"},
          "docType": {"type":"string"},
          "id": {"type":"string"},
          "userId": {"type":"string"},
          "visaType": {"type":"string"}
        }
    },
    "delete document":{
      "properties":{
        "docId":{"type":"string"},
        "email":{"type":"string"},
        "password":{"type":"string"}
      }
    },
    "post document":{
      "properties":{
        "email":{"type":"string"},
        "password":{"type":"string"},
        "docName":{"type":"string"},
        "docSerialNum":{"type":"string"},
        "docPublishedDate":{"type":"date"},
        "docExpiryDate":{"type":"date"},
        "docPublishOrg":{"type":"string"},
        "docType":{"type":"VISA or PASSPORT or DRIVERLICENSE"},
        "docDetailSerialNum":{"type":"string"}
      }
    },
    "get all document":{
      "properties":{
        "email":{"type":"string"}
      }
    },
    "get specific document":{
      "properties":{
        "email":{"type":"string"}
      }
    },
    "user modify":{
      "properties":{
        "password":{"type":"string"},
        "newPassword":{"type":"string"}
      }
    },
    "user login":{
      "properties":{
        "email":{"type":"string"},
        "password":{"type":"string"}
      }
    },
    "new user": {
      "properties":{
        "email": {"type":"string"},
        "password": {"type": "string"},
        "name": {"type":"string"},
        "birthday": {"type":"date"},
        "address": {"type":"string"},
        "sex": {"type":"string"},
        "nation": {"type":"string"}
      }
    },
    "user": {
      "properties": {
        "userid": {
          "type": "integer"
        },
        "email": {
          "type": "string"
        },
        "passwd": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "callnumber": {
          "type": "string"
        },
        "refresh": {
          "type": "string"
        },
        "salt": {
          "type": "string"
        }
      }
    }
  }
}